comp 310 assignment 1: fixup job handling involved in fg builtin
The fg builtin calls waitpid, which results in the waitpid being called
for a *second time* for the foregrounded process when check_jobs is run
before displaying the prompt. The upshot of this is that when waitpid is
called a second time, it returns -1 rather than new information
regarding the process, so the JobSpec would never get removed from the
job list. This is resolved by having the fg builtin set running = 0 on
the JobSpec of the process to foreground and by modifying the check_jobs
function to remove any jobs that have been marked as running = 0.