comp 520 assignment 1: restructure to conform to required format
All the code is moved into `src`, as required.
The cabal build script now compiles a `minilancheck` executable taking
one file name as a command-line argument, printing "Valid" if the
file parses and pringing "Invalid" otherwise.

Building and running instructions are written in README.
A name is proposed for the dragon in DragonName.txt.
A script `run` is added to first try to run a locally built
`minilangcheck` and fall back to a predistributed `minilangcheck` placed
in `src`.