jakemath: improvements
The mandatory argument for \range can be empty now, in which case the upper
bound is omitted. Hence, \range[5]{} shows up as {5, ...} in the output.
Much better than typing \{ 5, \ldots \} amirite? Also, this means that
\range{} shows up as {1, ...} because the starting value defaults to one.

A new command \nderiv{N}{X} is added to show the "Nth derivative with respect
to X" fraction.

The \evaluated[LO]{HI} command has been added. This shows up as a vertical bar
with sub- and superscripts.
If you only want a lower bound, e.g. for a subscript like "t=0", then use
\evaluated{LO}.

New probability distributions:
  * \HGeo for the hypergeometric distribution
  * \BetaD for the beta distribution. Although \Beta isn't a latex command
    (since capital beta is apparently indistinguishable from capital B), we
    append D for consistency with \GammaD.